Conflicts:
	library/Fisma/Menu.php

Conflict in Menu.php because Andrew was restructuring while I was changing privilege names. This was easy to resolve by taking his structure and then re-adding my privilege names.

There was also a surprising phpcs failure on Finding.php because a field named deleted_at was being referenced. I changed it to array access instead of object access as a quick fix.

+++ application/models/listener/IndexListener.php
@@ -95,9 +95,13 @@ class IndexListener extends Fisma_Doctrine_Record_Listener
             return;
         }
 
-        $record = $event->getInvoker();
+        $record   = $event->getInvoker();
+        $index    = new Fisma_Index(get_class($record));
+        $modified = $record->getLastModified();
 
-        $index = new Fisma_Index(get_class($record));
-        $index->delete($record);
+        // If the record is softDeleted, do nothing. Otherwise, delete the record from the index.
+        if (!in_array('deleted_at', $modified)) {
+            $index->delete($record);
+        }
     }
 }
